dissolve common module
