+++ common/Cargo.toml
@@ -12,4 +12,5 @@ description = "common traits and utility functions used by multiple tantivy subc
 byteorder = "1.4.3"
 
 [dev-dependencies]
+proptest = "1.0.0"
 rand = "0.8.4"
